requirements.txt
python-telegram-bot>=21.0.0
python-dotenv>=1.0.0
gspread>=5.0.0
oauth2client>=4.1.3
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
flask>=2.3.0

